+++ R/zzz.R
@@ -1,2 +1,4 @@
 #' @importFrom rlang abort warn %||%
+#' @importFrom utils hasName
+#' @importFrom R6 R6Class
 NULL